third_party/blink/web_tests/platform/generic/fast/loader/for-window-event-onload-scripts-expected.txt

ALERT: This script SHOULD be run because it has both 'for=window' and 'event=onload'.
ALERT: This script SHOULD be run because it has both 'for=window' and 'event=onload()'.
ALERT: This script SHOULD be run because it only has a for attribute.
ALERT: This script SHOULD be run because it only has an event attribute.
ALERT: This script SHOULD be run because it has both 'for=   window   ' and 'event=onload()' and the whitespace should be irrelevant.
ALERT: This script SHOULD be run because it has both 'for=window' and 'event=   onload()   ' and the whitespace should be irrelevant.
ALERT: This script SHOULD be run because it has both 'for=   window   ' and 'event=   onload()   ' and the whitespace should be irrelevant.
ALERT: This script SHOULD be run because it has both 'for=   window   ' and 'event=onload' and the whitespace should be irrelevant.
ALERT: This script SHOULD be run because it has both 'for=window' and 'event=   onload   ' and the whitespace should be irrelevant.
ALERT: This script SHOULD be run because it has both 'for=   window   ' and 'event=   onload   ' and the whitespace should be irrelevant.
